Implement new method of saving dataset
CurrentTick is a counter from 0 to TicksWithPause-1 and is incremented each game tick, returning to 0 each round. At CurrentTick == 1, the game is paused until TicksWithPause-2 is reached. At that point, screen content and objects content will be synchronized: a screenshot can be safely taken and bounding rects will be aligned to the screenshot. At TicksWithPause-1 the game is unpaused, letting the game progress for a tick. Then the cycle starts over.